requirements.txt
Flask==3.0.3
google-cloud-firestore==2.17.0
gunicorn==23.0.0